rename ContactBook2 component
