+++ src/frontend/web_application/src/modules/routing/components/RoutingProvider.jsx
@@ -23,7 +23,7 @@ import RemoteIdentitySettings from '../../../scenes/RemoteIdentitySettings';
 import ApplicationSettings from '../../../scenes/ApplicationSettings';
 import Tags from '../../../scenes/TagsSettings';
 import Discussion from '../../../scenes/Discussion';
-import ContactBook from '../../../scenes/ContactBook2';
+import ContactBook from '../../../scenes/ContactBook';
 import PageNotFound from '../../../scenes/PageNotFound';
 import DevicesSettings from '../../../scenes/DevicesSettings';
 import NewDeviceInfo from '../../../scenes/NewDeviceInfo';
