--- src/frontend/web_application/src/scenes/ContactBook/components/ContactFilters/style.scss
@@ -1,24 +1,0 @@
-@import '../../../../styles/common';
-@import '../../../../styles/vendor/bootstrap_foundation-sites';
-
-.m-contacts-filters {
-  @include flex-grid-row($size: expand)
-
-  &__item {
-    @include flex-grid-row;
-    align-items: center;
-    margin-top: $co-component__spacing;
-
-    @include breakpoint(medium) {
-      margin-top: 0;
-    }
-  }
-
-  &__label {
-    @include flex-grid-column(shrink);
-  }
-
-  &__select {
-    @include flex-grid-size;
-  }
-}
