+++ src/frontend/web_application/src/scenes/ContactBook/components/ContactItem/style.scss
