+++ src/frontend/web_application/src/scenes/ContactBook/components/TagList/style.scss
