+++ src/frontend/web_application/src/scenes/ContactBook/contact-book-menu.scss
