--- src/frontend/web_application/src/scenes/ContactBook2/components/ContactList/index.js
@@ -1,16 +1,0 @@
-import { createSelector } from 'reselect';
-import { connect } from 'react-redux';
-import Presenter from './presenter';
-import { settingsSelector } from '../../../../store/selectors/settings';
-
-const mapStateToProps = createSelector(
-  [settingsSelector],
-  ({
-    contact_display_order: contactDisplayOrder, contact_display_format: contactDisplayFormat,
-  }) => ({
-    contactDisplayOrder,
-    contactDisplayFormat,
-  })
-);
-
-export default connect(mapStateToProps)(Presenter);
