--- src/frontend/web_application/src/scenes/ContactBook2/style.scss
@@ -1,44 +1,0 @@
-@import '../../styles/common';
-@import '../../styles/vendor/bootstrap_foundation-sites';
-
-.s-contact-book {
-  &__sidebar {
-    padding-right: $co-margin;
-    padding-left: $co-margin;
-  }
-
-  &__tags-title {
-    margin: 20px 0 12px 0;
-    color: #505050;
-    font-size: 18px;
-    line-height: 14px;
-  }
-
-  &__import {
-    @include flex-grid-row;
-    margin-top: 12px;
-    margin-bottom: $co-margin;
-  }
-
-  &__action-button {
-    text-transform: uppercase;
-  }
-
-  &__select-all {
-    display: inline-block;
-    vertical-align: middle;
-    margin-left: 10px
-  }
-
-  &__load-more {
-    padding-top: $co-margin;
-    padding-bottom: $co-margin;
-    text-align: center;
-  }
-
-  @include breakpoint(medium) {
-    &__sidebar {
-      @include flex-grid-column(shrink);
-    }
-  }
-}
